Extracted a test to use its own actor system
