+++ amqp/src/test/scala/akka/stream/alpakka/amqp/scaladsl/AmqpConnectionProvidersSpec.scala
@@ -6,22 +6,10 @@ package akka.stream.alpakka.amqp.scaladsl
 
 import java.net.ConnectException
 
-import java.util.concurrent.atomic.AtomicInteger
-
-import akka.dispatch.ExecutionContexts
 import akka.stream.alpakka.amqp._
-import akka.stream.scaladsl.Source
-import akka.util.ByteString
-import com.rabbitmq.client._
-
-import scala.concurrent.{Future, Promise}
-import scala.concurrent.duration._
+import com.rabbitmq.client.ConnectionFactory
 
 class AmqpConnectionProvidersSpec extends AmqpSpec {
-
-  override implicit val patienceConfig = PatienceConfig(10.seconds)
-  private implicit val executionContext = ExecutionContexts.sameThreadExecutionContext
-
   "The AMQP Default Connection Providers" should {
     "create a new connection per invocation of LocalAmqpConnection" in {
       val connectionProvider = AmqpLocalConnectionProvider
@@ -90,7 +78,7 @@ class AmqpConnectionProvidersSpec extends AmqpSpec {
     }
   }
 
-  "The AMQP Cached Connection Provider with automatic release" should {
+  "The AMQP Reusable Connection Provider with automatic release" should {
     "reuse the same connection from LocalAmqpConnection and release it when the last client disconnects" in {
       val connectionProvider = AmqpLocalConnectionProvider
       val reusableConnectionProvider = AmqpCachedConnectionProvider(connectionProvider)
@@ -147,58 +135,9 @@ class AmqpConnectionProvidersSpec extends AmqpSpec {
       connection1.isOpen should be(false)
       connection2.isOpen should be(false)
     }
-
-    "registers and unregisters a single connection shutdown hook per graph" in {
-      val spyConnectionPromise = Promise[Connection]
-      val shutdownsAdded = new AtomicInteger()
-      val shutdownsRemoved = new AtomicInteger()
-
-      val connectionFactory = new ConnectionFactory()
-      val connectionProvider = DelegatingConnectionProvider(
-        AmqpConnectionFactoryConnectionProvider(connectionFactory, List(("localhost", 5672))),
-        conn => {
-          val proxy = new AmqpProxyConnection(conn) {
-            override def addShutdownListener(shutdownListener: ShutdownListener): Unit = {
-              shutdownsAdded.incrementAndGet()
-              super.addShutdownListener(shutdownListener)
-            }
-
-            override def removeShutdownListener(shutdownListener: ShutdownListener): Unit = {
-              shutdownsRemoved.incrementAndGet()
-              super.removeShutdownListener(shutdownListener)
-            }
-          }
-
-          spyConnectionPromise.success(proxy)
-          proxy
-        }
-      )
-
-      val reusableConnectionProvider =
-        AmqpCachedConnectionProvider(provider = connectionProvider, automaticRelease = true)
-
-      def queueName = "amqp-conn-prov-it-spec-simple-queue-" + System.currentTimeMillis()
-      val queueDeclaration = QueueDeclaration(queueName)
-
-      val amqpSink = AmqpSink.simple(
-        AmqpSinkSettings(reusableConnectionProvider)
-          .withRoutingKey(queueName)
-          .withDeclarations(queueDeclaration)
-      )
-
-      val input = Vector("one", "two", "three", "four")
-
-      Future
-        .traverse(input)(in => Source.single(ByteString(in)).runWith(amqpSink))
-        .flatMap(_ => spyConnectionPromise.future)
-        .futureValue
-
-      shutdownsAdded.get() should equal(input.size)
-      shutdownsRemoved.get() should equal(input.size)
-    }
   }
 
-  "The AMQP Cached Connection Provider without automatic release" should {
+  "The AMQP Reusable Connection Provider without automatic release" should {
     "reuse the same connection from LocalAmqpConnection" in {
       val connectionProvider = AmqpLocalConnectionProvider
       val reusableConnectionProvider = AmqpCachedConnectionProvider(connectionProvider, automaticRelease = false)
